Added reading BMP, but only INFOHEADER == 40 and biBitCount == 24

+++ src/DebugConsole.cpp
@@ -168,7 +168,7 @@ namespace con
 				}
 				case 'f':
 				{
-					double value = va_arg(args, double);
+					float value = va_arg(args, double);
 					int numDigit = 6;
 					float tempValue = value;
 					while (tempValue > 1)
